Feat(SyncNotes): Sincroniza as anotações entre os dispositivos logados
