+++ src/services/note/NoteService.ts
@@ -15,6 +15,7 @@ import DinoAgentService from '../dino_agent/DinoAgentService'
 import NoteSyncLocalStorage from './local_storage/NoteSyncLocalStorage'
 import DinoAgentStatus from '../../types/dino_agent/DinoAgentStatus'
 import NoteUpdateModel from '../../types/note/NoteUpdateModel'
+import NoteModel from '../../types/note/NoteModel'
 
 class NoteService {
   //#region GET
@@ -448,6 +449,48 @@ class NoteService {
     }
 
     NoteSyncLocalStorage.setShouldSync(true)
+  }
+
+  updateNotesFromServer = async (newVersion: number) => {
+    const localVersion = this.getVersion()
+
+    if (newVersion > localVersion) {
+      const request = DinoAgentService.get(DinoAPIURLConstants.NOTE_GET)
+
+      if (request.status === DinoAgentStatus.OK) {
+        try {
+          const response = await request.get()
+
+          if (response.status === HttpStatus.OK) {
+            const notes: NoteModel[] = response.body
+
+            const noteDocs: NoteDoc[] = notes.map(
+              (n) =>
+                ({
+                  external_id: n.id,
+                  order: n.order,
+                  answer: n.answer,
+                  answered: n.answered,
+                  lastUpdate: n.lastUpdate,
+                  question: n.question,
+                  tagNames: n.tags,
+                  savedOnServer: true,
+                } as NoteDoc)
+            )
+
+            this.setVersion(newVersion)
+
+            await NoteDatabase.removeAll()
+            await NoteDatabase.putAll(noteDocs)
+
+            return
+          }
+        } catch {
+          /**TO-DO Salvar log de erro */
+        }
+      }
+      NoteSyncLocalStorage.setShouldSync(true)
+    }
   }
 
   //#endregion
