+++ src/types/note/NoteWebSocketAlertUpdateModel.ts
@@ -1,0 +1,3 @@
+export default interface NoteWebSocketAlertUpdateModel {
+  newVersion: number
+}
